@@ -18,48 +18,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/cdk": "^5.0.0-rc0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/material": "^5.0.0-rc0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
+    "@angular/animations": "^10.0.7",
+    "@angular/cdk": "^10.1.2",
+    "@angular/common": "^10.0.7",
+    "@angular/compiler": "^10.0.7",
+    "@angular/core": "^10.0.7",
+    "@angular/forms": "^10.0.7",
+    "@angular/http": "^7.2.16",
+    "@angular/material": "^10.1.2",
+    "@angular/platform-browser": "^10.0.7",
+    "@angular/platform-browser-dynamic": "^10.0.7",
+    "@angular/router": "^10.0.7",
     "@deck.gl/core": "^7.3.15",
     "@deck.gl/google-maps": "^7.3.15",
     "@deck.gl/layers": "^7.3.15",
     "@turf/bbox": "^6.0.1",
-    "@types/crypto-js": "^3.1.44",
+    "@types/crypto-js": "^3.1.47",
     "@types/d3-scale": "^1.0.11",
     "@types/gapi": "0.0.35",
-    "@types/googlemaps": "^3.38.0",
+    "@types/googlemaps": "^3.39.11",
     "angular-webstorage-service": "^1.0.2",
-    "codemirror": "^5.52.2",
+    "codemirror": "^5.56.0",
     "colorbrewer": "^1.1.0",
     "core-js": "^2.6.11",
     "crypto-js": "^4.0.0",
-    "d3-color": "^1.4.0",
+    "d3-color": "^1.4.1",
     "d3-scale": "^1.0.7",
-    "firebase": "^7.13.1",
-    "js-file-download": "^0.4.11",
+    "firebase": "^7.17.1",
+    "js-file-download": "^0.4.12",
     "ng2-codemirror": "^1.1.3",
     "ngx-color-picker": "^5.3.8",
-    "rxjs": "^5.5.2",
+    "rxjs": "^6.6.2",
     "serve": "^10.0.2",
     "zone.js": "^0.8.14"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/cli": "^7.3.10",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
+    "@angular-devkit/build-angular": "*",
+    "@angular/cli": "^10.0.5",
+    "@angular/compiler-cli": "^10.0.7",
+    "@angular/language-service": "^10.0.7",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "^2.0.8",
-    "@types/node": "^13.9.8",
+    "@types/node": "^13.13.15",
     "codelyzer": "~3.2.0",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
@@ -69,11 +69,11 @@
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "node-sass": "^4.12.0",
-    "protractor": "^5.4.3",
+    "node-sass": "^4.14.1",
+    "protractor": "^5.4.4",
     "ts-node": "~3.2.0",
     "tslint": "~5.7.0",
-    "typescript": "^3.8.3"
+    "typescript": "^3.9.7"
   },
   "browser": {
     "crypto": false
